add test and fix negative number
